Create individual_property_landing_page.cshtml
This is the new landing page for each individual property. This will contain all the photos and information added when submitting a property. Each property that was submitted by a Seller, Auctioneer etc. Should have a page like this. When you select the more info or details button on a property card, thats located on the home page or on the properties page. You will end up on this page. This is also where you have the "Bid on this Property" button. That should lead you to the auction room to bid on this property